Renames util function. Now uses imported functions from lib.
